Rename source files for naming to be consistent
