+++ examples/utility/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(window SHARED Window.c)
+add_library(window SHARED window.c)
 target_link_libraries(window PRIVATE SDL2)
 target_include_directories(
 	window
