+++ src/CMakeLists.txt
@@ -3,9 +3,9 @@ set(
 	context.c
 	buffer.c
 	triangle.c
-	Framebuffer.c
-	Type.c
-	Texture.c
+	framebuffer.c
+	type.c
+	texture.c
 	stb_image.c
 	vec.c
 	mat.c
