+++ src/rasterizer.h
@@ -2,5 +2,5 @@
 
 #include "context.h"
 #include "buffer.h"
-#include "Texture.h"
+#include "texture.h"
